fixed exportSilently
